Fix MacOS compilation stuff
Also add cast to print hex numbers, not chars

+++ source/src/FDC765.cc
@@ -46,7 +46,7 @@ void FDC765::clock() {
 #ifdef DEBUGFDC765
                 } else {
                     cout << "Command byte " << static_cast<uint32_t>(cmdIndex) << ": "
-                        << cmdBuffer[cmdIndex] << endl;
+                        << static_cast<uint32_t>(cmdBuffer[cmdIndex]) << endl;
 #endif
                 }
             }
